+++ db/migrate/20250329123939_add_actor_type_to_actors.rb
@@ -1,0 +1,5 @@
+class AddActorTypeToActors < ActiveRecord::Migration[7.2]
+  def change
+    add_column :federails_actors, :actor_type, :string, null: true
+  end
+end