+++ CHANGELOG.md
@@ -33,6 +33,7 @@ Breaking changes should be prefixed by `[**BREAKING**]` (without the quotes), to
 ### Added
 
 - `Federails::Configuration#open_registrations` now supports a proc in addition to booleans.
+- `Federails::Actor` now stores the actor's type (`actor_type`)
 
 ### Changed